Enhancements + new feature
Enhanced time elapsed from using times like HH:MM:SS (which were not always accurate), to a progress bar that uses events to track progress. Commented out most of the code related to the time strings, as it may be useful again in the future.

New feature: togglable percussion (most songs don't sound good with it turned on though).