Enhancements and Prepping
Enhanced progress bars to show up as a notification instead.

Preparing to switch over to a new web midi player library which has no issues playing midi on the website, unlike the current library that I'm currently using.

Afterwards, planning to move player buttons to inside of the elapsing progress notification. While, leaving the other mod options in the same spot... (but eventually also making them appear in a notification of their own)